# api/openapispec/codes/openapi.json
{"openapi":"3.0.1","info":{"title":"Dysr Codes Api","description":"Provides Codes Information.","version":"1.0.0"},"paths":{"/api/v1/codes":{"get":{"tags":["Codes Api"],"summary":"Get all User Codes.","operationId":"GetCodes-fm","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ModelsUserCode"}}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}}}},"post":{"tags":["Codes Api"],"summary":"Creates a UserCode.","operationId":"CreateCode-fm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCode-FmInput"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsUserCode"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}}}}},"/api/v1/codes/{id}":{"get":{"tags":["Codes Api"],"summary":"Get all User Codes.","operationId":"GetCode-fm","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsUserCode"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}}}},"put":{"tags":["Codes Api"],"summary":"Update a pet.","operationId":"UpdateCode-fm","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCode-FmInput"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsUserCode"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}}}},"delete":{"tags":["Codes Api"],"summary":"Deletes a UserCode.","operationId":"DeleteCode-fm","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}}}}},"/api/v1/healthcheck":{"get":{"tags":["Codes Api"],"summary":"Checks API is healthy.","operationId":"Healthcheck","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsHealthcheck"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsAPIError"}}}}}}}},"components":{"schemas":{"CreateCode-FmInput":{"type":"object","properties":{"key":{"type":"string"},"options":{"$ref":"#/components/schemas/ModelsOptions"},"value":{"type":"string"}}},"ModelsAPIError":{"type":"object","properties":{"message":{"type":"string","description":"The error message returned to the client.","example":"Pet not found."}}},"ModelsCode":{"type":"object","properties":{"key":{"type":"string"},"options":{"$ref":"#/components/schemas/ModelsOptions"},"value":{"type":"string"}}},"ModelsHealthcheck":{"type":"object","properties":{"message":{"type":"string","description":"The status of the API.","example":"The API is healthy"}}},"ModelsOptions":{"type":"object","properties":{"colorDark":{"type":"string"},"colorLight":{"type":"string"},"height":{"type":"integer","format":"int32"},"text":{"type":"string"},"width":{"type":"integer","format":"int32"}}},"ModelsUserCode":{"type":"object","properties":{"code":{"$ref":"#/components/schemas/ModelsCode"},"created":{"type":"integer","format":"int64"},"id":{"type":"string"},"options":{"$ref":"#/components/schemas/ModelsOptions"},"updated":{"type":"integer","format":"int64"},"user":{"type":"string"}}},"UpdateCode-FmInput":{"type":"object","properties":{"key":{"type":"string"},"options":{"$ref":"#/components/schemas/ModelsOptions"},"value":{"type":"string"}}}}},"tags":[{"name":"Codes Api","description":"Provides all codes infos."}]}